refactor: BLOAT_PREVENTION compliance - Extract components & remove JIRA dependencies
Major architectural cleanup following BLOAT_PREVENTION_SYSTEM.md:

COMPONENT EXTRACTION (DRY Compliance):
- Extract JiraClient → .claudedirector/lib/integration/jira_client.py
- Extract ConfigManager → .claudedirector/lib/config/jira_config.py
- Extract data models → .claudedirector/lib/core/models.py
- Reduce weekly_reporter.py: 1915→1697 lines (218 lines removed, 11.4% reduction)

RETROSPECTIVE SYSTEM - JIRA REMOVAL:
- Remove ChatEnhancedWeeklyReporter inheritance (JIRA-based)
- Create standalone PersonalRetrospectiveSystem (NO JIRA dependencies)
- Update documentation: emphasize personal reflection only
- Architecture: 3 questions, no business data, no project tracking

ARCHITECTURE COMPLIANCE:
- Single Responsibility Principle restored
- DRY violations eliminated across 8+ files
- PROJECT_STRUCTURE.md compliance validated
- Component centralization complete
