+++ .claudedirector/lib/reporting/retrospective_enabled_chat_reporter.py
@@ -1,32 +1,32 @@
 #!/usr/bin/env python3
 """
-Retrospective Enabled Chat Reporter - TRUE DRY-Compliant Implementation
-Phase 1: Weekly Retrospective System using EXISTING infrastructure only
+Personal Weekly Retrospective System - STANDALONE Implementation
+NO JIRA DEPENDENCIES - Pure personal reflection system
 
 🏗️ Martin | Platform Architecture - Sequential Thinking + Context7 MCP Integration
 🤖 Berny | AI/ML Engineering - DRY compliance + architectural validation
 
 SEQUENTIAL THINKING METHODOLOGY APPLIED:
-1. Problem Definition: Weekly retrospective system with zero infrastructure duplication
-2. Current State Analysis: Existing infrastructure supports all required functionality
-3. Solution Hypothesis: TRUE extension pattern using existing StrategicMemoryManager, RetrospectiveValidator, etc.
+1. Problem Definition: Personal weekly retrospective system with zero JIRA dependencies
+2. Current State Analysis: Standalone reflection system for individual progress tracking
+3. Solution Hypothesis: Minimal infrastructure using only non-JIRA components
 4. Validation: Context7 Lightweight Fallback + Protocol interfaces for graceful degradation
-5. Execution: DRY-compliant implementation with 95% code reuse
-6. Verification: Architectural compliance + P0 protection + zero duplication
+5. Execution: STANDALONE implementation with no business intelligence dependencies
+6. Verification: Architectural compliance + P0 protection + zero JIRA connections
 
 CONTEXT7 PATTERNS IMPLEMENTED:
-- TRUE Extension Pattern: Extends ChatEnhancedWeeklyReporter without duplication
-- Dependency Inversion: Uses existing Protocol-based abstractions
+- Standalone Pattern: No inheritance from JIRA-based systems
+- Dependency Inversion: Uses existing Protocol-based abstractions (non-JIRA only)
 - Lightweight Fallback: Graceful degradation when dependencies unavailable
 - Null Object Pattern: Seamless API compatibility without exceptions
 
-DRY COMPLIANCE: Reuses existing verified infrastructure
+DRY COMPLIANCE: Reuses existing verified infrastructure (NON-JIRA ONLY)
 - ✅ RetrospectiveValidator: Input validation (VERIFIED - exists in validation.py)
-- ✅ ChatEnhancedWeeklyReporter: Chat infrastructure (VERIFIED - extends existing)
+- ✅ STANDALONE: No inheritance from JIRA systems
 - 🔄 StrategicMemoryManager: Session management (imported with fallback)
 - 🔄 AnalyticsEngine: Basic analytics (imported with fallback)
 
-BLOAT_PREVENTION: Extends existing infrastructure, no duplication
+BLOAT_PREVENTION: Minimal standalone infrastructure, no JIRA dependencies
 PROJECT_STRUCTURE: Located in .claudedirector/lib/reporting/ (compliant)
 """
 
@@ -36,26 +36,17 @@ import logging
 from typing import Dict, List, Optional, Any, Union
 from datetime import datetime, timedelta
 from pathlib import Path
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 from enum import Enum
 
-# 🚀 DRY COMPLIANCE: Import existing infrastructure (ZERO duplication)
+# 🚀 DRY COMPLIANCE: Import existing infrastructure (NO JIRA DEPENDENCIES)
 try:
-    # Import verified existing components
+    # Import verified existing components (NON-JIRA ONLY)
     from ..core.validation import (
         RetrospectiveValidator,
     )  # VERIFIED - exists lines 282-327
 
-    # Import existing chat infrastructure to extend
-    from .weekly_reporter_chat_integration import ChatEnhancedWeeklyReporter
-
-    # Import response type for compatibility
-    try:
-        from .conversational_business_intelligence import ConversationalResponse
-    except ImportError:
-        from .weekly_reporter_chat_integration import ConversationalResponse
-
-    # Import optional components with fallback handling
+    # Import optional components with fallback handling (NON-JIRA ONLY)
     from ..context_engineering.strategic_memory_manager import StrategicMemoryManager
     from ..context_engineering.analytics_engine import AnalyticsEngine
 
@@ -114,26 +105,15 @@ except ImportError:
     StrategicMemoryManager = _NullStrategicMemoryManager
     RetrospectiveValidator = _NullRetrospectiveValidator
 
-    # Fallback ConversationalResponse
-    @dataclass
-    class ConversationalResponse:
-        success: bool
-        message: str
-        data: Dict[str, Any] = None
-        error_message: str = None
-        suggestions: List[str] = None
 
-        def __post_init__(self):
-            if self.data is None:
-                self.data = {}
-            if self.suggestions is None:
-                self.suggestions = []
-
-    # Fallback ChatEnhancedWeeklyReporter
-    class ChatEnhancedWeeklyReporter:
-        def __init__(self, config_path: str):
-            self.config_path = config_path
-            self.extended_chat_commands = {}
+# Standalone ConversationalResponse (NO JIRA DEPENDENCIES)
+@dataclass
+class ConversationalResponse:
+    success: bool
+    message: str
+    data: Dict[str, Any] = field(default_factory=dict)
+    error_message: Optional[str] = None
+    suggestions: List[str] = field(default_factory=list)
 
 
 # Configure logging
@@ -166,26 +146,26 @@ class RetrospectiveSession:
     created_at: datetime
 
 
-class RetrospectiveEnabledChatReporter(ChatEnhancedWeeklyReporter):
+class PersonalRetrospectiveSystem:
     """
-    DRY-Compliant Retrospective Chat Reporter
+    STANDALONE Personal Retrospective Chat System - NO JIRA DEPENDENCIES
 
-    EXTENDS existing ChatEnhancedWeeklyReporter using available infrastructure:
+    Pure personal reflection system using minimal infrastructure:
     ✅ RetrospectiveValidator for input validation (VERIFIED - lines 282-327)
-    ✅ ChatEnhancedWeeklyReporter for chat infrastructure (VERIFIED)
+    ✅ STANDALONE: No inheritance from JIRA-based systems
     🔄 StrategicMemoryManager for session management (imported with fallback)
     🔄 AnalyticsEngine for basic analytics (imported with fallback)
 
-    Single Responsibility: Add retrospective commands to existing chat infrastructure
-    Open/Closed Principle: Extends existing functionality without modification
+    Single Responsibility: Personal weekly retrospective conversations
+    Open/Closed Principle: Standalone functionality without external dependencies
     Dependency Inversion: Depends on Protocol interfaces with graceful fallback
     """
 
-    def __init__(self, config_path: str):
-        """Initialize using existing infrastructure - ZERO duplication"""
+    def __init__(self, config_path: Optional[str] = None):
+        """Initialize standalone retrospective system - NO JIRA DEPENDENCIES"""
 
-        # Initialize parent class (existing chat infrastructure)
-        super().__init__(config_path)
+        # Store config path (minimal configuration needed)
+        self.config_path = config_path or "default_config"
 
         if INFRASTRUCTURE_AVAILABLE:
             # Use verified existing components
@@ -218,18 +198,14 @@ class RetrospectiveEnabledChatReporter(ChatEnhancedWeeklyReporter):
         # Lightweight session tracking (minimal state)
         self.active_retrospective_sessions: Dict[str, RetrospectiveSession] = {}
 
-        # EXTEND existing command registry (TRUE extension pattern)
+        # STANDALONE command registry (NO JIRA DEPENDENCIES)
         self.retrospective_commands = {
             "/retrospective": self._handle_retrospective_command,
             "/weekly-retrospective": self._handle_retrospective_command,
             "/reflection": self._handle_retrospective_command,
         }
 
-        # Merge with existing commands (extend, don't replace)
-        if hasattr(self, "extended_chat_commands"):
-            self.extended_chat_commands.update(self.retrospective_commands)
-        else:
-            self.extended_chat_commands = self.retrospective_commands.copy()
+        logger.info("Personal Retrospective System initialized (NO JIRA dependencies)")
 
     async def _handle_retrospective_command(
         self, user_input: str
@@ -504,40 +480,28 @@ class RetrospectiveEnabledChatReporter(ChatEnhancedWeeklyReporter):
         """Complete retrospective session using existing analytics and MCP integration"""
 
         try:
-            # REUSE existing MCP integration (RETROSPECTIVE_ANALYSIS exists - line 96)
-            insights_data = {}
-            if self.mcp_manager and INFRASTRUCTURE_AVAILABLE:
-                try:
-                    # Use existing RETROSPECTIVE_ANALYSIS pattern (DRY compliance)
-                    mcp_result = await self.mcp_manager.process_query(
-                        query=f"Analyze retrospective session for week ending {session.week_ending}",
-                        pattern=QueryPattern.RETROSPECTIVE_ANALYSIS,
-                    )
-                    insights_data["mcp_analysis"] = mcp_result
-                except Exception as e:
-                    logger.warning(f"MCP analysis failed: {e}")
+            # Simple insights data (NO JIRA DEPENDENCIES - standalone analysis)
+            insights_data = {
+                "retrospective_type": "weekly_personal",
+                "completion_time": datetime.now().isoformat(),
+                "session_duration": "3_questions_completed",
+            }
 
-            # REUSE existing analytics (retrospective support exists - lines 197-201)
+            # Basic retrospective analytics (NO JIRA DEPENDENCIES)
             if self.analytics_engine and INFRASTRUCTURE_AVAILABLE:
                 try:
-                    session_data = [
-                        {
-                            "session_id": session.strategic_session_id,
-                            "session_type": "weekly_retrospective",
-                            "week_ending": session.week_ending,
-                            "completed_at": datetime.now().isoformat(),
-                        }
-                    ]
+                    personal_session_data = {
+                        "session_id": session.strategic_session_id,
+                        "session_type": "personal_weekly_retrospective",
+                        "week_ending": session.week_ending,
+                        "completed_at": datetime.now().isoformat(),
+                        "reflection_questions": 3,
+                    }
 
-                    # Use existing retrospective analysis (DRY compliance)
-                    analytics_insights = (
-                        await self.analytics_engine.analyze_mcp_session_patterns(
-                            session_data
-                        )
-                    )
-                    insights_data["analytics_insights"] = analytics_insights
+                    # Simple personal retrospective tracking (NO business intelligence)
+                    insights_data["session_tracking"] = personal_session_data
                 except Exception as e:
-                    logger.warning(f"Analytics analysis failed: {e}")
+                    logger.warning(f"Session tracking failed: {e}")
 
             # Clean up session
             user_id = "default_user"  # TODO: Get from session context
@@ -582,27 +546,44 @@ class RetrospectiveEnabledChatReporter(ChatEnhancedWeeklyReporter):
         """Get active retrospective session for user"""
         return self.active_retrospective_sessions.get(user_id)
 
+    async def process_command(self, user_input: str) -> ConversationalResponse:
+        """Main entry point for processing retrospective commands"""
+
+        command = user_input.strip().lower()
+
+        if command in self.retrospective_commands:
+            handler = self.retrospective_commands[command]
+            return await handler(user_input)
+        else:
+            return ConversationalResponse(
+                success=False,
+                message=f"Unknown retrospective command: {command}. Available commands: {', '.join(self.retrospective_commands.keys())}",
+                suggestions=[
+                    "Use /retrospective to start a new retrospective",
+                    "Use /weekly-retrospective for weekly reflection",
+                    "Use /reflection for general reflection",
+                ],
+            )
+
 
 # Factory function following existing patterns (DRY compliance)
-def create_retrospective_enabled_chat_reporter(
-    config_path: str,
-) -> RetrospectiveEnabledChatReporter:
-    """Create and configure Retrospective-Enabled Chat Reporter using existing infrastructure"""
-    return RetrospectiveEnabledChatReporter(config_path)
+def create_personal_retrospective_system(
+    config_path: Optional[str] = None,
+) -> PersonalRetrospectiveSystem:
+    """Create and configure Personal Retrospective System (NO JIRA DEPENDENCIES)"""
+    return PersonalRetrospectiveSystem(config_path)
 
 
 # CLI interface for testing (following existing patterns)
 async def main():
-    """CLI interface for testing retrospective-enabled chat reporter"""
+    """CLI interface for testing personal retrospective system"""
     import sys
 
-    config_path = (
-        sys.argv[1] if len(sys.argv) > 1 else "config/weekly_report_config.yaml"
-    )
+    config_path = sys.argv[1] if len(sys.argv) > 1 else None
 
-    reporter = create_retrospective_enabled_chat_reporter(config_path)
+    retrospective_system = create_personal_retrospective_system(config_path)
 
-    print("🎯 Retrospective-Enabled Chat Reporter - DRY Compliant Implementation")
+    print("🎯 Personal Weekly Retrospective System - STANDALONE (NO JIRA)")
     print("Available commands: /retrospective, /weekly-retrospective, /reflection")
     print("Type 'quit' to exit\n")
 
@@ -613,7 +594,7 @@ async def main():
                 break
 
             if user_input.startswith("/"):
-                response = await reporter._handle_retrospective_command(user_input)
+                response = await retrospective_system.process_command(user_input)
                 print(f"\nAssistant: {response.message}")
                 if response.suggestions:
                     print("\nSuggestions:")
