+++ docs/development/weekly-retrospective-system/tasks-phase1.md
@@ -10,10 +10,10 @@
 - **Action**: Add retrospective tables to existing schema
 - **Pattern**: Extends existing DatabaseManager schema system
 
-### ✅ 1.2 Command Integration - **REQUIRED**
-- **Component**: ChatEnhancedWeeklyReporter
-- **Action**: Add `/retrospective` command mapping
-- **Pattern**: Extends existing command infrastructure
+### ✅ 1.2 Standalone System - **REQUIRED**
+- **Component**: PersonalRetrospectiveSystem (STANDALONE - NO JIRA)
+- **Action**: Create standalone `/retrospective` command system
+- **Pattern**: Independent personal reflection system
 
 ### ✅ 1.3 Session Management - **REUSE EXISTING**
 - **Component**: StrategicMemoryManager (existing)
@@ -38,29 +38,32 @@
 ## Implementation Architecture
 
 ```python
-# DRY-COMPLIANT IMPLEMENTATION
+# STANDALONE IMPLEMENTATION (NO JIRA DEPENDENCIES)
 from ..context_engineering.strategic_memory_manager import StrategicMemoryManager
 from ..context_engineering.analytics_engine import AnalyticsEngine
 from ..core.validation import RetrospectiveValidator
-from ..mcp.mcp_integration_manager import QueryPattern
 
-class RetrospectiveEnabledChatReporter(ChatEnhancedWeeklyReporter):
-    def __init__(self, config_path: str):
-        super().__init__(config_path)
+class PersonalRetrospectiveSystem:
+    def __init__(self, config_path: Optional[str] = None):
+        # NO inheritance from JIRA systems - completely standalone
+        self.config_path = config_path or "default_config"
 
-        # REUSE existing infrastructure
+        # REUSE existing infrastructure (NON-JIRA ONLY)
         self.session_manager = StrategicMemoryManager()
         self.validator = RetrospectiveValidator()
 
-        # Add minimal command mapping
+        # STANDALONE command mapping (personal reflection only)
         self.retrospective_commands = {
             '/retrospective': self._handle_retrospective_command,
+            '/weekly-retrospective': self._handle_retrospective_command,
+            '/reflection': self._handle_retrospective_command,
         }
 ```
 
 ## Summary
 
-**Files to Modify**: 2 files only
-**New Infrastructure**: 0 components
-**DRY Compliance**: 100%
-**Implementation Size**: <50 lines total
+**Files Created**: 1 standalone file (retrospective_enabled_chat_reporter.py)
+**JIRA Dependencies**: 0 (completely standalone)
+**DRY Compliance**: 100% (reuses non-JIRA infrastructure only)
+**Implementation Size**: <200 lines total
+**Architecture**: Standalone personal reflection system with NO business intelligence dependencies
